p29.21 refactor: thread ids.ExpenseReportID + ids.ExpenseReportLineID everywhere
Typed-ID capstone, commit 4: the coupled expense-report / expense-report-line
ids are now distinct types end-to-end. Same shape as the budget pair (a PK pair
joined by one FK: ExpenseReportLine.ReportID -> ExpenseReport).

Threaded:
- sqlc: ExpenseReport.ID, ExpenseReportLine.{ID,ReportID}, ExpenseReportsVersion
  .EntityID, ExpenseReportLinesVersion.{EntityID,ReportID}; every report-id/
  line-id query param + return + version ID_2.
- store: all Create/Add/Update/Remove/Replace/Submit/Resubmit/Reject/Convert/
  Discard/Get/List signatures, transitionExpenseReport/loadExpenseReport/
  requireEditable, ExpenseReportLineDesired.ID (the by-id diff key -> the replace
  set's existing/kept maps are now line-id-keyed), the two version helpers'
  entityID params, and the Create* returns.
- web: expenses.go + expenses_review.go convert at the HTTP-parse seam
  (ids.ExpenseReportID(parseID)) and read typed store fields into int64 model/
  render values; loadOwnedReport keeps its int64 param, converts at the store call.
- cmd: expense-report reject CLI converts at the strconv boundary.
- synth: DemoIDs.{DraftReport,SubmittedReport,PostedReport} typed.

Boundary discipline (the traps the advisor flagged): ExpenseReport.PostedTransac-
tionID stays sql.NullInt64 (nullable txn FK -> TransactionID, pervasive, last);
ConvertExpenseReport's postedTxnID param + PostAndConvertExpenseReport's int64
RETURN are both TransactionID and stay int64 (the return looks like a report id
but is the posted txn id); submitterID (UserID), subsidiaryID, line account/fund/
program stay int64; *_versions.ID audit PKs stay int64. Generic AssertVersioned
stays int64; test call sites pass int64(reportID)/int64(lineID). Test-local SQL
helpers (expenseReportStatus/reviewNotes/postedTxnID) typed to ExpenseReportID.

make lint (0 issues) + make test (Go + 148 node) + make check green; expense
store/web/review + devseed + CLI tests pass; report goldens byte-identical.

+++ cmd/cuento/devseed_test.go
@@ -122,7 +122,7 @@ func TestSeedSampleExpenseReport(t *testing.T) {
 	reportID := submitted[0].ID
 	// The report went through the write funnel: created, then the submit appended an
 	// 'update' version (draft -> submitted), which is the latest op (rule 5).
-	testutil.AssertVersioned(t, f.DB, "expense_reports", reportID, "update")
+	testutil.AssertVersioned(t, f.DB, "expense_reports", int64(reportID), "update")
 
 	lines, err := f.Store.ExpenseReportLines(ctx, reportID)
 	if err != nil {
